requirements.txt
astroid==2.4.2
autopep8==1.5.3
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
Cython==0.29.20
demjson==2.2.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
pycodestyle==2.6.0
pytesseract==0.3.4
requests==2.23.0
six==1.15.0
termux-apt-repo==0.4
tesseract==0.1.3
toml==0.10.1
urllib3==1.25.9
wrapt==1.12.1
